+++ Sources/View/PhotoCollectionViewCell.swift
@@ -9,6 +9,5 @@
 import UIKit
 
 class PhotoCollectionViewCell: UICollectionViewCell {
-    
     @IBOutlet weak var photoImageView: UIImageView!
 }
